refactor: 💡 Updated the License API
Updated the spelling of License and standardized internal functions to
be more consistent

BREAKING CHANGE: 🧨 Changed the spelling of the License API from Licence to License.
Additionally the license.create method has been replaced with
license.add and has a required first parameter that is the license. The
license.all method now returns all licenses and no longer the license
for the current user. This is now done by license.show.