fix(core): Adding better handling for form type arguments
This should be able to account for any form arguments passed to the requestHelper.

BREAKING CHANGE: Trigger.pipeline now has a defined variables property to house all the pipeline variables to be set. Before these variables were set on the root object, but this made it difficult to parse any future options that may be added to this function.

fixes: #568