refactor: remove camelization from type system
Removes camelization from being represented in the type system
Updates BaseService to remove generic parameter
Updates types leveraging BaseService to remove generic camelize paramter

In general, it seems like the API can return different casings in different responses.
If that's the case, handling it when processing requests and responses seems to me to
be a better approach that letting it propagate throughout the type system.

This may be a breaking change, as if I am understanding it correctly, objects before
this pr had two properties: fooBar and foo_bar.

Hopefully, after this change, this will not be the case.