Add support for username lookup. (#2147)
Similar to Users.search() this looks up a username directly using
Users.username(). This is useful when you want to grab user information
for a known username without having to run a search which could be slow.

For example a common name such as `david` is slower with search than a
username lookup and also the first result is not for this username
meaning you have to parse through a large array rather than just accept
the first result.